:recycle: Move accommodation type selection to its own component.
Why? In preparation of re-using this component on a per-stop basis.
